+++ rollup.config.es.js
@@ -1,0 +1,7 @@
+const configBuilder = require("./rollup.config")
+const pkg = require("./package.json")
+
+export default configBuilder(
+  { file: pkg.module, format: "es", sourcemap: true },
+  { external: ["os", "tty", "util"] }
+)